// src/exchange/interfaces/interface.ws.ts
export interface Subscribe {
  type: string;
  user?: string;
  coin?: string;
  interval?: number;
}
export type SubscribeType =
  | 'allMids'
  | 'notification'
  | 'webData2'
  | 'candle'
  | 'l2Book'
  | 'userFills'
  | 'orderUpdates'
  | 'trades'
  | 'userEvents'
  | 'userFundings'
  | 'userNonFundingLedgerUpdates';